+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -10,6 +10,7 @@ Version 1.6.0
               inside any ftl.html file. (ra)
  * 2013-07-30 Direct streaming from assets folder without jetty reload in dev mode. Cool
               for developing js apps inside the assets folder. (ra)
+ * 2013-08-05 Fixed bug in OPTION method of routes. (ra)
  
 Version 1.5.1
 =============